A binary tree is a special type of tree in which every node or vertex has either no child node or one child node or two child nodes. A binary tree is an important class of a tree data structure in which a node can have at most two children.
Tree represents the nodes connected by edges. It is a non-linear data structure.
We create a tree data structure in python by using the concept OS node discussed earlier. 
We designate one node as root node and then add more nodes as child nodes.
To insert into a tree we use the same node class created above and add a insert class to it. 
The insert class compares the value of the node to the parent node and decides to add it as a left node or a right node.
 Finally the PrintTree class is used to print the tree.